refactor: (#56) features lobby 슬라이스 내 index.ts 배럴 파일 생성 및 import 경로 일괄 수정
