+++ src/features/lobby/index.ts
@@ -1,3 +1,5 @@
-export * from './constants';
-export * from './model';
-export * from './ui';
+export { RoomActionButtons } from './ui/RoomActionButtons';
+export { RoomDashboard } from './ui/RoomDashboard';
+export { TeamSection } from './ui/TeamSection';
+export { TeamTab } from './ui/TeamTab';
+export { useRoomUI } from './ui/useRoomUI';
