--- src/features/lobby/model/index.ts
@@ -1,2 +1,0 @@
-export * from './useRoom';
-export * from './roomSelectors';
