+++ src/features/lobby/ui/RoomActionButtons.tsx
@@ -1,4 +1,5 @@
 import type { ButtonHTMLAttributes } from 'react';
+
 import { cn } from '@/shared/lib/cn';
 
 interface RoomActionButtonsProps {
