--- src/features/lobby/ui/index.ts
@@ -1,7 +1,0 @@
-export * from './ScalableText';
-export * from './MenuButton';
-export * from './PlayerSlot';
-export * from './RoomActionButtons';
-export * from './RoomDashboard';
-export * from './TeamSection';
-export * from './TeamTab';
